+++ app/Http/Controllers/Api/V1/UserController.php
@@ -10,8 +10,6 @@ use App\Http\Controllers\Api\V1\ApiController;
 use App\Http\Requests\Api\V1\StoreUserRequest;
 use App\Http\Requests\Api\V1\UpdateUserRequest;
 use App\Http\Requests\Api\V1\ReplaceUserRequest;
-use Illuminate\Auth\Access\AuthorizationException;
-use Illuminate\Database\Eloquent\ModelNotFoundException;
 
 class UserController extends ApiController
 {
@@ -21,7 +19,7 @@ class UserController extends ApiController
      */
     public function index(AuthorFilter $filter)
     {
-       return UserResource::collection(User::filter($filter)->paginate());
+        return UserResource::collection(User::filter($filter)->paginate());
     }
 
     /**
@@ -29,12 +27,10 @@ class UserController extends ApiController
      */
     public function store(StoreUserRequest $request)
     {
-        try {
-            $this->isAble('store', User::class);
+        if ($this->isAble('store', User::class)) {
             return new UserResource(User::create($request->mappedAttribute()));
-        } catch (AuthorizationException $ex) {
-            return $this->error('You are not authorized to create the resource.', 401);
         }
+        return $this->notAuthorized('You are not authorized to create the resource.');
     }
 
     /**
@@ -51,58 +47,33 @@ class UserController extends ApiController
     /**
      * Update the specified resource in storage.
      */
-    public function update(UpdateUserRequest $request, $user_id)
+    public function update(UpdateUserRequest $request, User $user)
     {
-        try {
-            $user = User::FindOrFail($user_id);
-
-            $this->isAble('update', $user);
-
+        if ($this->isAble('update', $user)) {
             $user->update($request->mappedAttribute());
-
             return new UserResource($user);
-
-        } catch (ModelNotFoundException $exception) {
-            return $this->error('User cannot be found.', 404);
-            
-        } catch (AuthorizationException $ex) {
-            return $this->error('You are not authorized to update the resource.', 401);
         }
+        return $this->notAuthorized('You are not authorized to update the resource.');
     }
 
-    public function replace(ReplaceUserRequest $request, $user_id)
+    public function replace(ReplaceUserRequest $request, User $user)
     {
-        try {
-            $user = User::FindOrFail($user_id);
-
-            $this->isAble('replace', $user);
-
+        if ($this->isAble('replace', $user)) {
             $user->update($request->mappedAttribute());
-
             return new UserResource($user);
-        } catch (ModelNotFoundException $exception) {
-            return $this->error('User cannot be found.', 404);
-        } catch (AuthorizationException $ex) {
-            return $this->error('You are not authorized to replace the resource.', 401);
         }
+        return $this->notAuthorized('You are not authorized to replace the resource.');
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy($user_id)
+    public function destroy(User $user)
     {
-        try {
-            $user = User::findOrFail($user_id);
-
-            $this->isAble('delete', $user);
-
+        if ($this->isAble('delete', $user)) {
             $user->delete();
             return $this->ok('User deleted successfully.');
-        } catch (ModelNotFoundException $exception) {
-            return $this->error('User cannot be found.', 404);
-        } catch (AuthorizationException $ex) {
-            return $this->error('You are not authorized to delete the resource.', 401);
         }
+        return $this->notAuthorized('You are not authorized to delete the resource.');
     }
 }
